MOVEONLY: Chainparams: Separate globals and functions depending on them
