+++ src/globals/chainparamsbaseglobals.cpp
@@ -1,0 +1,10 @@
+// Copyright (c) 2010 Satoshi Nakamoto
+// Copyright (c) 2009-2014 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include "globals/chainparamsbaseglobals.h"
+
+#include "chainparamsbase.h"
+
+Container<CBaseChainParams> cGlobalChainBaseParams;
