+++ src/globals/chainparamsbaseglobals.h
@@ -1,0 +1,16 @@
+// Copyright (c) 2014 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_GLOBALS_CHAINPARAMSBASE_H
+#define BITCOIN_GLOBALS_CHAINPARAMSBASE_H
+
+#include "templates.hpp"
+
+#include <string>
+
+class CBaseChainParams;
+
+extern Container<CBaseChainParams> cGlobalChainBaseParams;
+
+#endif // BITCOIN_GLOBALS_CHAINPARAMSBASE_H
